requirements.txt
streamlit
pillow
pytesseract
pdf2image
langchain-community
langchain
langchain-groq
langchain-core
sentence-transformers
faiss-cpu
gradio
